--- src/Web3RequiredAlert.css
@@ -1,22 +1,0 @@
-.Web3RequiredAlert {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  padding: 0.5em;
-  padding: 1em;
-  font-size: 2em;
-  background: #fff;
-  color: #222;
-  text-align: center;
-  line-height: 1.6em;
-}
-
-.Web3RequiredAlert p {
-  max-width: 700px;
-  margin: auto;
-}
