--- src/components/AccountHeader.css
@@ -1,51 +1,0 @@
-.AccountHeader {
-  display: flex;
-  min-height: 105px;
-}
-
-.AccountHeader .Identicon {
-  width: 60px;
-  height: 60px;
-}
-
-.AccountHeader .image {
-  width: 70px; /* take account margin, because of flex */
-  height: 60px;
-  background: #ddd;
-  margin-right: 10px;
-}
-
-.AccountHeader .Content {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  padding: 0.1em 0;
-  justify-content: center;
-  overflow-x: auto;
-}
-
-.AccountHeader .Header {
-  font-size: 0.9em;
-}
-
-.AccountHeader .SubHeader {
-  font-size: 1.4em;
-  white-space: nowrap;
-  overflow-x: auto;
-  padding: 0.2em;
-}
-
-.AccountHeader .SubHeader .Name {
-  display: inline-block;
-  margin-right: 10px;
-}
-
-.AccountHeader .SubHeader .Country {
-  display: inline-block;
-}
-
-@media only screen and (max-width: 767px) {
-  .AccountHeader .SubHeader {
-    overflow-x: auto;
-  }
-}
