--- src/components/ApplicationContainer.css
@@ -1,11 +1,0 @@
-.ApplicationContainer {
-
-}
-
-.ApplicationContainer > .grid > .column:first-child {
-  padding-right: 0;
-}
-
-:root:root .ApplicationContainer .PublisherApplicationInfo {
-  margin-bottom: 1rem;
-}
