--- src/components/ClaimRewardContainer.css
@@ -1,3 +1,0 @@
-.ClaimRewardContainer {
-
-}
