--- src/components/DomainClaimRewardInProgressContainer.css
@@ -1,12 +1,0 @@
-.DomainClaimRewardInProgressContainer {
-  background: rgba(255,255,255,0.9);
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  padding: 2em;
-}
