--- src/components/PublisherApplicationForm.css
@@ -1,16 +1,0 @@
-.PublisherApplicationForm {
-  position: relative;
-}
-
-
-.PublisherApplicationForm .header {
-  border-bottom: 4px solid #4585c7;
-}
-
-:root:root .PublisherApplicationForm .form label {
-  text-transform: uppercase;
-}
-
-.PublisherApplicationForm .icon.external {
-  vertical-align: super;
-}
