--- src/components/Trollbox.css
@@ -1,85 +1,0 @@
-.TrollboxContainer {
-  display: block;
-  margin: 0;
-  padding: 0;
-  width: 100%;
-}
-
-.TrollboxHeader {
-  font-weight: bold;
-  padding-bottom: 0.2em;
-  display: none;
-}
-
-.TrollboxMessages {
-  display: block;
-  width: 100%;
-  height: 268px;
-  overflow-x: hidden;
-  overflow-y: auto;
-  padding: 0.3em;
-  color: #686972;
-  font-size: 12px;  
-  font-weight: 400;
-  min-height: 13px;
-  font-family: Open Sans,sans-serif;
-}
-.TrollboxMessages strong{
-  color:grey;
-}
-
-.TrollboxMessages::-webkit-scrollbar {
-  width: 5px;
-  height: 5px;
-}
-
-.TrollboxMessages::-webkit-scrollbar-track {
-  background: #fff;
-}
-
-.TrollboxMessages::-webkit-scrollbar-thumb {
-  background-color: #ccc;
-  border: 1px solid #ccc;
-}
-
-.TrollboxMessages::-webkit-scrollbar-thumb:hover {
-  background-color: #aaa;
-}
-
-.TrollboxMessagesList {
-  width: 100%;
-  height: 100%;
-  list-style: none;
-  margin: 0;
-  padding: 0;
-}
-
-.TrollboxMessagesList li {
-  line-height: 1.2;
-}
-
-.TrollboxMessage {
-  display: block;
-  width: 100%;
-}
-
-.TrollboxForm {
-  padding-top: 0.2em;
-  display: block;
-  width: 100%;
-}
-
-.TrollboxInput {
-  display: block;
-  width: 100%;
-  padding: .67857143em 1em;
-  font-size: 1em;
-  background: #F4F7FA;
-  color: #686972;
-  border: 1px solid transparent;
-}
-
-.TrollboxInput:focus {
-  outline: 0;
-  border-color: #536DEB;
-}
