+++ src/components/dashboard/ClaimRewardContainer.js
