+++ src/components/dashboard/DomainClaimRewardInProgressContainer.js
