+++ src/components/sidebar/PublisherApplicationFormInProgress.css
