+++ src/components/sidebar/PublisherApplicationFormInProgress.js
